// src/stats/types/league-dash-player-stats.ts
export interface LeagueDashPlayerStatsParams {
  lastNGames?: number;
  measureType?: string;
  month?: number;
  opponentTeamID?: number;
  paceAdjust?: string;
  perMode?: string;
  period?: number;
  plusMinus?: string;
  rank?: string;
  season?: string;
  seasonType?: string;
  college?: string;
  conference?: string;
  country?: string;
  dateFrom?: string;
  dateTo?: string;
  division?: string;
  draftPick?: string;
  draftYear?: number;
  gameScope?: string;
  gameSegment?: string;
  height?: string;
  leagueID?: string;
  location?: string;
  outcome?: string;
  pORound?: number;
  playerExperience?: string;
  playerPosition?: string;
  seasonSegment?: string;
  shotClockRange?: number;
  starterBench?: string;
  teamID?: number;
  twoWay?: string;
  vsConference?: string;
  vsDivision?: string;
  weight?: string;
}

export interface LeagueDashPlayerStatsRow {
  playerId: number;
  playerName: string;
  teamId: number;
  teamAbbreviation: string;
  age: number;
  gp: number;
  w: number;
  l: number;
  wPct: number | null;
  min: number;
  fgm: number;
  fga: number;
  fgPct: number | null;
  fg3m: number;
  fg3a: number;
  fg3Pct: number | null;
  ftm: number;
  fta: number;
  ftPct: number | null;
  oreb: number;
  dreb: number;
  reb: number;
  ast: number;
  tov: number;
  stl: number;
  blk: number;
  blka: number;
  pf: number;
  pfd: number;
  pts: number;
  plusMinus: number;
  nbaFantasyPts: number;
  dd2: number;
  td3: number;
  gpRank: number;
  wRank: number;
  lRank: number;
  wPctRank: number;
  minRank: number;
  fgmRank: number;
  fgaRank: number;
  fgPctRank: number;
  fg3mRank: number;
  fg3aRank: number;
  fg3PctRank: number;
  ftmRank: number;
  ftaRank: number;
  ftPctRank: number;
  orebRank: number;
  drebRank: number;
  rebRank: number;
  astRank: number;
  tovRank: number;
  stlRank: number;
  blkRank: number;
  blkaRank: number;
  pfRank: number;
  pfdRank: number;
  ptsRank: number;
  plusMinusRank: number;
  nbaFantasyPtsRank: number;
  dd2Rank: number;
  td3Rank: number;
  cfid: number;
  cfparams: unknown;
}

export interface LeagueDashPlayerStatsResponse {
  leagueDashPlayerStats: LeagueDashPlayerStatsRow[];
}